requirements.txt
appdirs==1.4.4
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
contourpy==1.1.0
cycler==0.11.0
Django==4.2.4
django-debug-toolbar==4.2.0
fonttools==4.42.0
frozendict==2.3.8
html5lib==1.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.3
matplotlib==3.7.2
multitasking==0.0.11
numpy==1.25.2
packaging==23.1
pandas==2.0.3
pandas-datareader==0.10.0
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
webencodings==0.5.1
yfinance==0.2.27
